AES: Added Test and Changed how you Set Keys
I added test for reading and writing from the
fifos as well as the entire encrypt and decrypt,
asserting that the flags are properly set and
what not. I also changed it such that you no
longer provide a key for every operation and the
KeyStatus is stored in the AES struct.